+++ domain/src/main/java/com/tistory/domain/usecase/AddStockUsecase.kt
@@ -6,15 +6,15 @@ import com.tistory.domain.repository.StockWithDividendRepository
 
 class AddStockUsecase(
     private val stockWithDividendRepository: StockWithDividendRepository
-) : SuspendUseCase<Pair<String, Int>, Any>() {
+) : SuspendUseCase<Pair<String, Float>, Any>() {
 
-    override suspend fun onSuccess(params: Pair<String, Int>) {
+    override suspend fun onSuccess(params: Pair<String, Float>) {
         val (ticker, stockCnt) = params
         stockWithDividendRepository.fetchAndPutStock(ticker, stockCnt)
         stockWithDividendRepository.fetchAndPutDividends(ticker)
     }
 
-    suspend fun get(ticker: String, stockCnt: Int, listener: BaseListener<Any>) {
+    suspend fun get(ticker: String, stockCnt: Float, listener: BaseListener<Any>) {
         build(Pair(ticker, stockCnt), listener)
     }
 }
